+++ IntegorTelegramBotListeningServices/ApiAggregation/Aggregators/StandardTelegramBotLongPollingAggregator.cs
@@ -26,7 +26,7 @@ namespace IntegorTelegramBotListeningServices.ApiAggregation.Aggregators
 			_dataServiceConfiguration = dataServiceOptions.Value;
 		}
 
-		public async Task AggregateAsync(IEnumerable<TelegramUpdateDto> updates, int botId)
+		public async Task AggregateAsync(IEnumerable<TelegramUpdateDto> updates, long botId)
 		{
 			Uri uri = new Uri(_dataServiceConfiguration.Url);
 			uri = new Uri(uri, $"telegram-events/aggregate-many/{botId}");
